requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python-headless==4.8.0.76
onnxruntime==1.15.1
pillow==10.0.0